restructured tests folder
