+++ test/widget_tests/home/flower_petal_widget_test.dart
@@ -7,7 +7,7 @@ import 'package:learning_compass_exp/store/app_state.dart';
 import 'package:learning_compass_exp/screens/home/widgets/flower/flower_petal.dart';
 import 'package:learning_compass_exp/store/reducers/app_state_reducer.dart';
 
-import '../mock_data.dart';
+import '../../mock_data.dart';
 
 void main() {
   final TestWidgetsFlutterBinding binding =
