+++ test/widget_tests/home/flower_widget_test.dart
@@ -1,8 +1,7 @@
+import 'dart:math';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:redux/redux.dart';
-import 'package:flutter_redux/flutter_redux.dart';
-import 'dart:math';
 
 import 'package:learning_compass_exp/screens/home/widgets/flower/flower.dart';
 import 'package:learning_compass_exp/screens/home/widgets/flower/flower_menu.dart';
@@ -11,7 +10,7 @@ import 'package:learning_compass_exp/screens/home/widgets/flower/flower_petal.da
 import 'package:learning_compass_exp/store/app_state.dart';
 import 'package:learning_compass_exp/store/reducers/app_state_reducer.dart';
 
-import '../mock_data.dart';
+import '../../mock_data.dart';
 
 void main() {
   final TestWidgetsFlutterBinding binding =
