+++ test/widget_tests/home/petal_icon_button_widget_test.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:redux/redux.dart';
-import 'package:flutter_redux/flutter_redux.dart';
 
 import 'package:learning_compass_exp/store/app_state.dart';
 import 'package:learning_compass_exp/app.dart';
@@ -11,7 +10,7 @@ import 'package:learning_compass_exp/store/reducers/app_state_reducer.dart';
 import 'package:learning_compass_exp/common/widgets/custom_icons_icons.dart';
 import 'package:learning_compass_exp/data/models/petal.dart';
 
-import '../mock_data.dart';
+import '../../mock_data.dart';
 
 void main() {
   final TestWidgetsFlutterBinding binding =
